pyproject.toml
[project]
name = "knowledge-arena"
version = "1.1.0"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "celery[redis]>=5.4",
    "redis>=5.0",
    "httpx>=0.27",
    "openai>=1.0",
    "bcrypt>=4.0",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "uuid7>=0.1",
    "websockets>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
include = ["app*", "tests*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
